% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_primes.R
\name{get_primes}
\alias{get_primes}
\title{Get primes below number}
\usage{
get_primes(upper_limit)
}
\arguments{
\item{upper_limit}{Number to which to generate primes.}
}
\value{
Vector of primes numbers below given limit (also included if prime).
}
\description{
Get primes using simple Sieve of Eratosthenes algorithm.
}
\examples{
get_primes(100)
}
\keyword{primes}
