% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greatest_common_div.R
\name{greatest_common_div}
\alias{greatest_common_div}
\title{Greates common divisor}
\usage{
greatest_common_div(a, b)
}
\arguments{
\item{a, b}{numbers.}
}
\value{
A number that is the greatest common divisor of \code{a} & \code{b}.
}
\description{
Gets greatest common divisor of two numbers.
}
\examples{
greatest_common_div(27,30)
}
\keyword{gcd}
