% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_hexagonal.R
\name{is_hexagonal}
\alias{is_hexagonal}
\title{Hexagonal number test}
\usage{
is_hexagonal(h)
}
\arguments{
\item{h}{A number.}
}
\value{
\code{TRUE} if hexagonal, \code{FALSE} if not.
}
\description{
Checks if number is hexagonal, based on formula Hn = n(2n−1).
}
\examples{
is_hexagonal(15)
is_hexagonal(20)
}
