% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_palindromic.R
\name{is_palindromic}
\alias{is_palindromic}
\title{Palindromic number}
\usage{
is_palindromic(number)
}
\arguments{
\item{number}{A number.}
}
\value{
\code{TRUE} if palindromic, \code{FALSE} if not.
}
\description{
Checks if given number is a palidrome, i.e., it can be read the same both ways.
}
\examples{
is_palindromic(12321)
is_palindromic(1231)
}
\keyword{primes}
