% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pandigital.R
\name{is_pandigital}
\alias{is_pandigital}
\title{Pandigital number test}
\usage{
is_pandigital(n, zero = F)
}
\arguments{
\item{n}{A number.}

\item{zero}{logical; to indicate whether zeros are to be included. Default to \code{FALSE}.}

\item{unique}{logical; to indicate whether repeated numbers are allowed. Default to \code{TRUE}.}
}
\value{
\code{TRUE} if pandigital, \code{FALSE} if not.
}
\description{
Checks if given number is pandigital, i.e., it makes use of all the digits from 1 to d exactly once, being d the number digits in \code{n}.
}
\examples{
is_pandigital(45321)
is_pandigital(103254, zero = T)
}
\keyword{pandigital}
