% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pentagonal.R
\name{is_pentagonal}
\alias{is_pentagonal}
\title{Pentagonal number test}
\usage{
is_pentagonal(p)
}
\arguments{
\item{p}{A number.}
}
\value{
\code{TRUE} if pentagonal, \code{FALSE} if not.
}
\description{
Checks if number is pentagonal, based on formula Pn = n(3n−1)/2.
}
\examples{
is_pentagonal(12)
is_pentagonal(15)
}
