% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_prime.R
\name{is_prime}
\alias{is_prime}
\title{Primality test}
\usage{
is_prime(n)
}
\arguments{
\item{n}{A number.}
}
\value{
\code{TRUE} if prime, \code{FALSE} if not.
}
\description{
Checks if given number is prime.
}
\examples{
is_prime(5)
is_prime(10)
}
\keyword{primes}
