% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_triangular.R
\name{is_triangular}
\alias{is_triangular}
\title{Triangular number test}
\usage{
is_triangular(m)
}
\arguments{
\item{m}{A number.}
}
\value{
\code{TRUE} if triangular, \code{FALSE} if not.
}
\description{
Checks if number is triangular, based on formula Tn = n*(n+1)/2.
}
\examples{
is_triangular(10)
is_triangular(20)
}
