% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_permutation.R
\name{nth_permutation}
\alias{nth_permutation}
\title{Nth Lexicographic Permutation}
\usage{
nth_permutation(vector, nth)
}
\arguments{
\item{vector}{A vector.}

\item{nth}{A number indicating the requested nth permutation of \code{vector}.}
}
\value{
The nth permutation of the given vector.
}
\description{
Gets the nth lexicographic permutation of a vector. Adaptation of Knuth's algorithm for generating all permutations (The Art of Computer Programming, V4, 7.2.1.2).
}
\examples{
v <- c(1,2,3)
nth_permutation(v, 4)
}
\keyword{permutations}
