% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_to_vector.R
\name{number_to_vector}
\alias{number_to_vector}
\title{Number to Vector}
\usage{
number_to_vector(number)
}
\arguments{
\item{number}{Number to parse into numerical vector.}
}
\value{
Numerical vector
}
\description{
Converts number into numerical vector, separating each digit as elements of the vector.
}
\examples{
vector_to_number(123)
}
