% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\title{Lexicographic permutations}
\usage{
permute(vector)
}
\arguments{
\item{vector}{A vector of elements to permute.}
}
\value{
Vector of permutations in lexicographic order.
}
\description{
Gets all permutations of given vector in lexicographic order using Knuth's algorithm (The Art of Computer Programming, V4, 7.2.1.2).
}
\examples{
v <- c(1,2,3)
permute(v)
}
\keyword{permutation}
