% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_factors.R
\name{prime_factors}
\alias{prime_factors}
\title{Prime Factors}
\usage{
prime_factors(number, primes)
}
\arguments{
\item{number}{A number to be factorized.}

\item{primes}{Optional. A vector of prime numbers to be used in the factorization. This makes the process faster, avoiding having to generate the first primes one by one. If not given, an initial vector of primes below 100 will be used as first input; the rest of primes will be generated by this algorithnm as needed.}
}
\value{
Vector of primes factors.
}
\description{
Gets prime factors of given number using a list of prime numbers.
}
\examples{
primes <- get_primes(100)
prime_factors(primes, 100)
}
\keyword{primes}
