% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proper_divisors.R
\name{proper_divisors}
\alias{proper_divisors}
\title{Proper divisors}
\usage{
proper_divisors(number)
}
\arguments{
\item{number}{A number.}
}
\value{
An ordered vector of proper divisors.
}
\description{
Gets proper divisors of a number n, i.e. numbers below n that divide evenly into n.
}
\examples{
proper_divisors(100)
}
