% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reversed_number.R
\name{reversed_number}
\alias{reversed_number}
\title{Reverse of a number}
\usage{
reversed_number(number)
}
\arguments{
\item{number}{A number.}
}
\value{
Reverse of number.
}
\description{
Return the reverse of a given number, i.e. same number but read from right to left.
}
\examples{
reversed_number(1423)
}
