% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\title{Swap positions of elements in vector}
\usage{
swap(vector, id_1, id_2)
}
\arguments{
\item{vector}{A vector of values.}

\item{id_1, id_2}{Indexes of elements in \code{vector} to be swapped.}
}
\value{
Vector with positions changed.
}
\description{
Swaps positions of elements in given vector.
}
\examples{
v <- c(1,2,3)
swap(v, 1, 3)
}
