% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_number.R
\name{vector_to_number}
\alias{vector_to_number}
\title{Vector to Number}
\usage{
vector_to_number(numeric_vector)
}
\arguments{
\item{numeric_vector}{Numerical vector to parse into number.}
}
\value{
Number
}
\description{
Converts numerical vector to a number, concatenating the numbers in given vector as consecutive digits of the number.
}
\examples{
vector_to_number(c(1,2,3))
}
